attempting to fix build issues
Adding the jekyll-responsive-image plugin causes build issues
on gitthub actioncs, because it requires rmagick which in turn
requires ImageMagick 6. The latest version, v7, which is installed
by on the github jekyll docker container is not compatible with rmagick
and therefor the build fails. We fix this by using another docker image
that does include the correct version of ImageMagick.
See: https://ineed.coffee/post/jekyll-docker-image-with-imagemagick6-instead-of-imagemagick7